pyproject.toml
[tool.poetry]
name = "fun-game"
version = "0.1.0"
description = ""
authors = ["Chris Busillo <info@shinycomputers.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.0.0"
cairosvg = "^2.7.1"
faker = "^19.3.1"
numpy = "^1.25.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
